+++ R/poisson_reg_glm.R
@@ -1,0 +1,9 @@
+#' Poisson regression via glm
+#'
+#' [stats::glm()] uses maximum likelihood to fit a model for count data.
+#'
+#' @includeRmd man/rmd/poisson_reg_glm.Rmd details
+#'
+#' @name details_poisson_reg_glm
+#' @keywords internal
+NULL