+++ R/poisson_reg_stan.R
@@ -1,0 +1,10 @@
+#' Poisson regression via stan
+#'
+#' [rstanarm::stan_glm()] uses Bayesian estimation to fit a model for
+#' count data.
+#'
+#' @includeRmd man/rmd/poisson_reg_stan.Rmd details
+#'
+#' @name details_poisson_reg_stan
+#' @keywords internal
+NULL