+++ NEWS.md
@@ -1,4 +1,4 @@
-# poissonreg (development version)
+# poissonreg 0.1.1
 
 * A default engine of `glm` was added for `poisson_reg()`.